server/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Auth
PyJWT>=2.10.0
bcrypt>=4.2.0

# Async tasks
celery[redis]>=5.4.0
redis>=5.2.0

# File handling
python-multipart>=0.0.18
aiofiles>=24.1.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0

# Dev tools
ruff>=0.8.0
